% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_list}
\alias{country_list}
\title{List of matrices grouped by country}
\format{
An object of class \code{list} of length 18.
}
\usage{
data(country_list)
}
\description{
A named list where each element corresponds to a country and contains
all matrices for that country.
}
\keyword{datasets}
