% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_matrices}
\alias{list_matrices}
\title{List all available SSA social contact matrices}
\usage{
list_matrices()
}
\value{
A tibble with one row per matrix and metadata columns.
}
\description{
This function returns a tibble containing metadata for all
social contact matrices included in the ssamixr package.
}
\examples{
list_matrices()
}
