% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssa_matrices_meta}
\alias{ssa_matrices_meta}
\title{Metadata for all SSA social contact matrices}
\format{
A tibble with columns:
\describe{
\item{matrix_id}{Matrix identifier}
\item{study_id}{Study identifier}
\item{author_year}{Author-year label}
\item{country}{Country name}
\item{location_type}{Location type}
\item{row_age_groups}{Row age groups}
\item{col_age_groups}{Column age groups}
}
}
\usage{
data(ssa_matrices_meta)
}
\description{
A tibble containing metadata for all matrices included in the package.
}
\keyword{datasets}
