% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plot_matrix_gradient}
\alias{plot_matrix_gradient}
\title{Plot a single social contact matrix (gradientn palette)}
\usage{
plot_matrix_gradient(
  m,
  blank_color = "grey90",
  value_text = TRUE,
  size = 4.5,
  col_label_orientation = "auto",
  base_size = 16
)
}
\arguments{
\item{m}{A list returned by get_matrix().}

\item{blank_color}{Color for missing cells.}

\item{value_text}{Whether to print numeric values in cells.}

\item{size}{Text size for cell values.}

\item{col_label_orientation}{"normal", "vertical", or "auto".}

\item{base_size}{Base font size for the plot.}
}
\value{
A ggplot object.
}
\description{
Plot a single social contact matrix (gradientn palette)
}
